Change back to orignal linked list data structure
The new linked list data structure is unfriendly for
sorting and leads to bad performance. Now changes back
to the original. In order to be compatible with the
use of concurrent merging, we add some functions for
adding and removing sentinel nodes. Thus we can use
these functions before or after concurrent merging.